refactored and document the randomization methods
