+++ lib/naer_services/XmlElementHandler/handle_xml_elements.dart
@@ -140,4 +140,17 @@ class XmlElementHandler {
       }
     }
   }
+
+  /// Replaces the text content of the specified XML element with the given new text.
+  ///
+  /// This function clears the existing children of the XML element and adds
+  /// a new text node with the specified [newText].
+  ///
+  /// Parameters:
+  /// - [element]: The XML element whose text content is to be replaced.
+  /// - [newText]: The new text to set for the XML element.
+  static void replaceTextInXmlElement(xml.XmlElement element, String newText) {
+    element.children.clear();
+    element.children.add(xml.XmlText(newText));
+  }
 }
