+++ lib/nier_cli/nier_cli_fork_utils/utils/check_paths.dart
@@ -1,4 +1,7 @@
+import 'package:NAER/naer_utils/change_tracker.dart';
 import 'package:NAER/nier_cli/nier_cli_fork_utils/utils/exception.dart';
+import 'package:NAER/nier_cli/nier_cli_fork_utils/utils/log_print.dart';
+import 'package:path/path.dart' as path;
 
 /// Retrieves the sorted enemies path from the list of arguments.
 ///
@@ -31,4 +34,28 @@ void validatePaths(String? sortedEnemiesPath, String? output) {
   if (output == null) {
     throw const FileHandlingException("Output path not specified");
   }
+}
+
+/// Asynchronously retrieves the file path to the mod metadata.
+///
+/// This function ensures that the settings directory exists, then constructs
+/// the path to the mod metadata file ('mod_metadata.json') within the 'ModPackage'
+/// directory. The path is logged for reference and returned as a string.
+///
+/// Returns:
+/// A `Future<String>` that completes with the file path to the mod metadata.
+Future<String> getMetaDataPath() async {
+  // Ensure the settings directory exists and get its path
+  final String settingsDirectoryPath =
+      await FileChange.ensureSettingsDirectory();
+
+  // Construct the path to the mod metadata file
+  final String metadataPath =
+      path.join(settingsDirectoryPath, 'ModPackage', 'mod_metadata.json');
+
+  // Log the found metadata path
+  logAndPrint("Found metadata at $metadataPath");
+
+  // Return the constructed metadata path
+  return metadataPath;
 }
